added include/aug.h, include/aug_plugin.h.
started the first draft of the plugin api.
essentially the idea is for a plugin to be
a shared object with at least an aug_plugin_init
and an aug_plugin_free function. the plugin
can register for callbacks at important IO
and screen events and control the screen
using the provided core api (in the form
of a struct with a bunch of function
pointers).