screen window api calls
added api calls to aug.h so that a plugin can allocate
space on the screen to present data to the user (for
example: a status bar at the bottom of the screen). also
added api calls so a plugin can allocate an ncurses panel
on top of the screen to interact with the user.