arbitrary size/location of terminal window
working on making sure the term_win code works
regardless of the size or location of the terminal
WINDOW * on the screen.

fixed a bug where a 'set window size' ioctl was
getting executed when the term structure still had
an invalid (=0) master pty, thus sending a SIGWINCH
to the parent process and messing everything up.