implemented key extension binding api calls
added a keymap.c module (and a TAP style test for it). the
module is a wrapper around a CCAN avl binary tree and keeps
track of mapping a key like 'r' to a callback function + user
pointer. still to be done: implement core functionality for a
user prefix for the extensions to actually work.