fixed dead lock bug in aug.c
when term_win is resized in screen module, the vterm terminal
is also resized. in some cases, a resize of the vterm terminal
causes updates to the screen which will in turn cause an callback
to the plugins. the cell update and move cursor callback functions
in aug.c assume that a lock is held on all the main structures, not
just screen and term. since the winch handler was only locking the
screen and term resources, we were unlocking some mutexes more than
once.