fixed child thread join hang bug
when a plugin attempted to kill a child process
created for a terminal, the master pty associated with
that terminal was not being found as readable by
the select call in child_io_loop and thus when
the plugin joined the IO loop thread, it was getting
stuck. this isnt exactly fixed, as i have no idea why
this happens, but opening a file before forking the child
(so both processes share the descriptor) fixes the issue.
obviously i have no idea why this works.