fixed nasty use after free bug in region map
region_map_keyregs_clear would free a key_regs pointer
and then subsequently allocate a new AVL tree to be
the new key_regs object. however since the input
parameter was a single pointer, the newly allocated
AVL tree was getting assigned to that local variable
and was not getting back to the caller. changed the
input parameter to be a double pointer so we can
replace the old AVL tree for real.