turn off some valgrind targets on OSX 10.5
there are some problems with using valgrind on OSX 10.5,
mainly that it crashes when sigwait is used. we can comment
out the call to sigwait in aug.c but this then causes valgrind
to report errors which i beleive are a result of not properly
processing signals. anyhow i think its best to just not use
valgrind on OSX 10.5 so the makefile disables valgrind targets
that test the full aug program (which uses sigwait) when 10.5 is
detected. im not sure if valgrind will work better on newer
versions of OSX, hopefully it will.