changed aug_plugin.h to provide a more simple interface
now global API objects are allocated on behalf of the
plugin and the aug_plugin_init symbol is implemented.
a new prototype aug_plugin_start is defined, for which
the plugin should define the implementation.

plugins which do not want such a simplified interface
(which comes with the baggage of secret global objects
and macros) can simply use the aug_plugin_definitions.h
include which is the same as what aug_plugin.h was
before.